more WIP refactoring / moving things
